RAG bot streaming
This PR adds streaming functionality to the RAG bot. The answer is streamed directly from the Ollama backend to the RAG micro service and to the UCE backend, there the message object is updated in realtime. Users can then poll messages using another endpoint to receive updates, this has been implemented for the UCE web ui already.

Open tasks:
- Implement streaming for OpenAI API and other local models/backends
- Integration of websockets into UCE to enable streaming without polling